+++ src/person.cpp
@@ -89,6 +89,8 @@ bool Person::reciprocal_infection_boolean(const Parameters &pars)
         
         for(auto m : m_gametocytogenic_strains){
           current_contribution = m_active_strains[m].relative_contribution(pars);
+          
+          // TODO: Add here the infectivity function to scale if ArtR
           if (current_contribution > highest_contribution){
             highest_contribution = current_contribution;
           }
@@ -110,7 +112,7 @@ bool Person::reciprocal_infection_boolean(const Parameters &pars)
     case SUBPATENT:
       return(rbernoulli1(pars.g_cU*highest_contribution));
     case TREATED:
-      return(rbernoulli1(m_cA*highest_contribution));
+      return(rbernoulli1(pars.g_cT*highest_contribution));
     case PROPHYLAXIS:
       return(false);
     default: